requirements.txt
django==3.2.7
djangorestframework==3.12.4
django-filter==2.4.0
markdown==3.3.4
requests==2.26.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
coverage==5.5
